+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Changelog
 
 ### [Unreleased]
-* 
+* Fix for object Update with null values
 
 ### [1.5.1] - 2017-05-11
 * Include xml documentation file and fix IDataStore method summaries